Move common function into common.h
